Moar eigen Eigenvalues.cpp: Deleted!
git-svn-id: http://theory.cm.utexas.edu/svn/eon@499 9d962951-508e-48a8-94b5-974f9cfda47c

+++ client_eigen/potentials/Aluminum/Aluminum.h
@@ -22,14 +22,14 @@ extern "C"
 
 /** Aluminum potential.*/
 class Aluminum : public PotentialsInterface
-{    
+{     
 public:
     Aluminum(void);
 	
     // To satify interface
     void initialize(void);    
     void cleanMemory(void);    
-    void force(long N, const double *R, const long *atomicNrs, double *F, double *U, const double *box);
+    void force(long N, const double *R, const int *atomicNrs, double *F, double *U, const double *box);
 };
 #endif
