+++ client_eigen/potentials/VASP/VASP.h
@@ -11,11 +11,11 @@ class VASP : public PotentialsInterface
 		~VASP();
         void initialize() {};
         void cleanMemory(void);    
-        void force(long N, const double *R, const long *atomicNrs, double *F, double *U, const double *box);
+        void force(long N, const double *R, const int *atomicNrs, double *F, double *U, const double *box);
 
 
     private:
-        void writeNEWCAR(long N, const double *R, long const *atomicNrs, const double *box);
+        void writeNEWCAR(long N, const double *R, int const *atomicNrs, const double *box);
         void readFU(long N, double *F, double *U);
         static long vaspRunCount;
 		static bool vaspRunning;
